Functions and files split, comments added.
